% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_hhmmss.R
\name{convert_hhmmss}
\alias{convert_hhmmss}
\title{Convert hh:mm:ss to hours}
\usage{
convert_hhmmss(hhmmss, forward = TRUE)
}
\arguments{
\item{hhmmss}{Either a string with "hh:mm:ss" or a decimal number of hours}

\item{forward}{Default is TRUE, thus performs the hh:mm:ss conversion to decimal. FALSE would do the opposite conversion.}
}
\description{
Convert hh:mm:ss to hours
}
\examples{
convert_hhmmss("07:56:08") / convert_hhmmss("09:46:22")
0.8040201 * convert_hhmmss("28:23:41")
convert_hhmmss(23.09548, forward=FALSE)
}
\keyword{biohazardCleanUp}
