% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_folders.R
\name{create_folders}
\alias{create_folders}
\title{Create Folders}
\usage{
create_folders(source, copy_path = FALSE)
}
\arguments{
\item{source}{The directory you wish to copy folders from.}

\item{copy_path}{Optional. The path where you want the copy to go to. Otherwise, it will just go to the directory above your source directory.}
}
\description{
Eventually, this should work for several different functionalities.
}
\examples{
create_folders(getwd())
}
\keyword{biohazardCleanUp}
